Refactor partitioning code.
